[KK-62] удалил mockexchangers, добавил изменения в основные файлы
